Move ParsedFeed to reader._parser. #307
